Fix Extensions.WebRequest.ProxyDecision metric
The metric was added in crrev.com/c/6102618, but the corresponding
code that figures out the reason for the decision was wrong, and
it always says -ForExtension regardless of the actual reason.

This fixes the unexpected reason override, and update the metric
name to have a suffix 2.

Bug: 385032014
Change-Id: I697ccbd05650baa6e1af30f1443278422be0ee17
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6188742
Commit-Queue: Takashi Toyoshima <toyoshim@chromium.org>
Auto-Submit: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1410739}