[Blink] Add variant of CreateResourceProviderForVideoFrame()
This CL adds a CreateResourceProviderForVideoFrame() variant that takes
in "native" SharedImage types. This will facilitate the removal of the
current variant that takes in SkImageInfo.

Note: At the current time we need to convert the passed-in
SharedImageFormat back to SkColorType to when creating the resource
providers. That will shortly change, so I'm creating this API with its
final shape rather than needing to do another later pass converting it
and its callers from SkColorType to SharedImageFormat.

Bug: 371227617
Change-Id: I5cb5065394019d3e467cea47696b54db68c9e820
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6194120
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1410763}