[functions] Handle omitted single parameter
We incorrectly enter the "consume argument" branch due to
first_parameter being 'true', but we shouldn't do that if nothing
is actually provided. We already consume leading whitespace, so we
can fix this issue by just checking if we're at the end.

Bug: 325504770
Change-Id: I733c08c8488faf8056a2f931b8f8bbef39e0d1a6
Fixed: 390581527
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6187335
Reviewed-by: Steinar H Gunderson <sesse@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1410788}