[ios] Add a test hook to not purge the data for discarded sessions
Some EG tests are flaky because the OS invokes the application
delegate's method -application:didDiscardSceneSessions: with a
list of session identifiers that contains identifiers of scene
that are active and connected, and that are still used after
the app is restarted.

Since the test app is re-installed when needed, and the tests
take care of clearing the data, there is no issue with never
purging the data when a window is discarded. Add a test hook
to never purge the data for discarded sessions for EG tests.

Bug: 390108895
Change-Id: Ia13dcf293bc993815068786277ae4fe96f39d589
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6196738
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Commit-Queue: Justin Cohen <justincohen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1410806}