[memory-infra] Report memory mappings count on Linux-based systems
On Linux, there is a kernel-enforced limit to the number of Virtual
Memory Areas (VMAs) that a process is allowed to have, with the default
limit (DEFAULT_MAX_MAP_COUNT in the kernel) set to a bit less than 2^16.

We are crash reports suggesting that Chromium is occasionally hitting
the limit. To see how close we are, this CL adds a per-process histogram
to report the number of VMAs.

Bug: 385400561
Change-Id: Id345840a182588ac51eeb809436bfaaff912b24b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6148597
Reviewed-by: Francois Pierre Doray <fdoray@chromium.org>
Reviewed-by: Joe Mason <joenotcharles@google.com>
Commit-Queue: Benoit Lize <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1410811}