Get rid of AutofillType constructors in FormFieldDataAndroid.
The one thing not perfect about this CL is that we're storing
FieldTypes::computed_type as std::string. The reason for that is that
HtmlFieldType has other string representation than FieldType (with a
prefix of HTML_TYPE - see [0]).

I think that the ultimate solution to this problem would be to inspect
the interfaces for the Java part - now we're passing heuristic_type,
server_type and computed_type and the only way for the Java code to know
computed type is from autocomplete is to parse computed_type.

However, I deem it far beyond the scope of this cl to rework the Java
part.

[0]
https://source.chromium.org/chromium/chromium/src/+/main:components/autofill/core/common/html_field_types.cc;l=10;drc=27d34700b83f381c62e3a348de2e6dfdc08364b8

Bug: 389018075

Change-Id: Ia18395a0af44062c05ecf40278bdf337d3eaaef4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6172467
Commit-Queue: Piotr Domaradzki <doland@google.com>
Reviewed-by: Florian Leimgruber <fleimgruber@google.com>
Reviewed-by: Jan Keitel <jkeitel@google.com>
Cr-Commit-Position: refs/heads/main@{#1410839}