Add feature parameter for switching button placement
We plan to run a multi-arm experiment when showing warnings on
suspicious notifications. The default behaviour has the "Unsubscribe"
button as the primary. In this CL, we've added a new parameter which -
when true - has "Unsubscribe" as the secondary and the other buttons
("Show notification" and "Always allow") as the primary.

Bug: 382475678
Change-Id: I3bf2fef6cd9825a3d371df74bc734c6479ab2d65
Low-Coverage-Reason: COVERAGE_UNDERREPORTED
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6192021
Reviewed-by: Peter Beverloo <peter@chromium.org>
Commit-Queue: Sarah Krakowiak <skrakowi@chromium.org>
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1410864}