[AutofillAi] Trigger new suggestion generation.
This Cl does the following:

1. Removes HasDataStored method in favor of GetSuggestionsV2.
This new method handles generating suggestions for the
new project iteration.
2. Updates BAM to use GetSuggestionsV2.
3. Makes the necessary test updates and refactoring for the CL
to build.

Notes:

This CL deletes a lot of code. Most of it is due to updating the
conditions such as form eligibility, which lead to old
callers being deleted, which led to orphan methods and so on. The
way things stand it is not too easy to right small CLs without working
around the old implementation and likely overcomplicating the code.
This CL does not intent to delete all code from the previous iteration,
only what it immediately touches/affects.

Bug: 389629573
Change-Id: I10096193ea0c51de4e5a6f9c8e4d05d9d28b91d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6187391
Reviewed-by: Jan Keitel <jkeitel@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Bruno Braga <brunobraga@google.com>
Cr-Commit-Position: refs/heads/main@{#1410879}