Remove outdated presubmit error for AlertDialog.Builder in VR
Before this patch, presubmit blocks changes that introduce new uses of
AlertDialog.Builder with a message that it won't work in VR. This
presubmit is outdated, because the VR mode to which it is referring to
has been removed.

This patch removes the outdated presubmit check.

After this patch, presubmit does not block changes that introduce new
uses of AlertDialog.Builder.

Bug: 391928689
Change-Id: If8617cef8c3798cd62b7d0297ee23462c36707b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6191455
Reviewed-by: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Rouslan Solomakhin <rouslan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1410910}