Prototyping "Inline Integrity"
This CL takes a first pass at implementing one spelling of the proposal
in https://mikewest.github.io/inline-integrity/. It only addresses
`<script>` elements; if we decide that it's a reasonable direction,
adding support for `<style>` should be pretty straightforward. If we
decide it's not a reasonable direction, this is easy to rip out. :)

Bug: 391907163
Change-Id: I35887e1c6d017ffbf3ea99f11bbd12f131c785e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6192623
Reviewed-by: Antonio Sartori <antoniosartori@chromium.org>
Reviewed-by: Fredrik Söderquist <fs@opera.com>
Commit-Queue: Mike West <mkwst@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1410927}