[system tracing] Use crashpad in the elevated tracing service
A crashpad process is spawned once the identity of the calling process
is determined. Crashes are only uploaded if consent has been granted to
the browser. The crash database is stored in a per-user subdirectory of
SystemTemp, and is deleted upon browser uninstall.

Bug: 41497361
Change-Id: I3a985010797b8bbe9de113dc9270f2904f798fd6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6157297
Commit-Queue: Avi Drissman <avi@chromium.org>
Auto-Submit: Greg Thompson <grt@chromium.org>
Reviewed-by: Will Harris <wfh@chromium.org>
Reviewed-by: Mark Mentovai <mark@chromium.org>
Reviewed-by: S Ganesh <ganesh@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1410944}