Update `IpProtectionCore` to store MdlType
The core object for IPP is instantiated per `NetworkContext`. In order
to correctly check if requests require IPP, the core must be aware of
what version of the MDL (incognito vs regular) to check. This change
adds a state stored in the core to keep track of which MDL to check in
for requests.

This change simply adds the mechanisms required to store the MDL type
state directly within the IPP core object. It should not currently be
set. This should not cause any functional change in the logic. A
followup cl will be written that correctly initializes the core from
the `NetworkContext`.

Change-Id: I0559e647916d4a48d41d59c45840419193d512a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6187189
Reviewed-by: Dustin Mitchell <djmitche@chromium.org>
Commit-Queue: Abhi Patel <abhipatel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1410946}