Add a retry loop around sqlite open for busy database
This CL is adding a loop around the call to sqlite3_open_v2(...)
for the case where the returned error code is kBusy.

The kBusy error code is the most common ones received from
our metrics. Since Chrome is not opening the file multiple times,
it should never happened. We are assuming it's an interaction
with third-party and are hoping that the retry loop will reduces
the amount of failures.

The metrics 'Sql.Database.Success.SqliteOpenRetries' is added to
track whether or not this approach is working.

Bug: 369635654
Change-Id: If83f5cbd98f81e6319b4f284c5d7c040fccd9644
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6191568
Reviewed-by: Anthony Vallée-Dubois <anthonyvd@chromium.org>
Reviewed-by: Greg Thompson <grt@chromium.org>
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1410947}