Add TextAutosizing to UKM
If we discover that vast majority of TextAutosizing occurs on google
properties, we can delete most of it because autosizing is only used to
detect the os font scale factor. That would fix most of the compat bugs
in Blink>TextAutosize in one fell swoop.

But if autosizing is broadly used, this will let us learn about the
real-world usage.

Bug: 391990606
Change-Id: I81f5510d2f3ac66f605052f60d2ae89d5184171e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6196162
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Commit-Queue: David Grogan <dgrogan@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1410948}