[Test Automation] WebContentsInteractionTestUtil to use polymorphism
This was a many-headed class that used the presence of various
parameters and members to determine if it was a tab web contents, a
standalone WebView, or something else when determining various
behaviors. This is a problem better solved with polymorphism than a
bunch of if and case blocks.

This CL splits the util into a primary interface class with most of
the functionality, and specializations for browser tabs and non-tab
WebContents.

This will make it easier to handle things like embedded <webview>
tags later.

Change-Id: Ie678f28c8b31890f0012f200da7390a77b2ab672
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6190714
Reviewed-by: Gavin Williams <gavinwill@chromium.org>
Reviewed-by: Christopher Grant <cjgrant@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1410997}