Implement accessibilityDocument as a stub implementation
Exposing the top-level URL on the NSWindow requires the URL to be made
available to the platform accessibility code. See crrev.com/c/6187085
for a candidate approach to doing so.

In the meantime, provide a stub implementation of this new API, along
with a corresponding test that will be updated once the implementation
details have been agreed upon.

Test: `out/.../browser_tests \
   --gtest_filter="*.AccessibilityDocumentExposedOnWindow*"`

Include-Ci-Only-Tests: true
Cq-Include-Trybots: luci.chromium.try:mac-fieldtrial-tester

Bug: 363275809
Change-Id: I933f6b53c228526b7d2b8e8f530bdff4a8aca943
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6163433
Commit-Queue: Joanmarie Diggs <jdiggs@igalia.com>
Reviewed-by: David Tseng <dtseng@chromium.org>
Reviewed-by: Jayson Adams <shrike@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1411013}