webnn: handle avgpool2d roundingType=ceil on CoreML
The output size is calculated by upstream caller based on the
roundingType, but CoreML needs to know the roundingType and it needs to
match with the output size.
When cache directory is not allowed access, a mismatch just silently
fail, but when CoreML has cache access it crashes on a size mismatch.

So reconstruct the roundingType from the output size.

Added one test case because currently all paddings are asymmetric but
coreml only support symmetric paddings for ceil mode.

Change-Id: Ic1771ccfae07d43bac36fb617ca2b9b9cb364ec2
Bug: 334914466
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6182746
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Commit-Queue: Phillis Tang <phillis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1411014}