[Test Automation] Add <webcontents> instrumentation
This expands WebContentsInteractionTestUtil to handle <webview> tags
that manifest as "inner frames" in HTML documents, such as the glic
window.

It creates a new private implementation class for this case which
watches a "parent" or outer instrumented web contents and creates the
inner instrumented contents when both are fully loaded.

This is *very* hard to test because <webview> only works in a very
limited number of cases and can't be simulated in a generic page. So
instead, glic will be tested using it, ensuring it works as expected.

As a later follow-up, we will determine how to write more generalized
tests, but that will take significant time and research.

Change-Id: I8a5665b0f7efde6d962882267123b49f65eedee8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6190715
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: William Liu <liuwilliam@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1411019}