Move interactive test API to use new inner webcontents logic
This refactors the interactive glic test to use the new inner
WebContents instrumentation introduced in Kombucha to avoid having to
use unreliable JS injection through the chrome.webviewTag API.

It also opens up mock glic to be used freely in interactive tests with
the full Kombucha API and not just a few one-off injection methods that
previously had to be hand-rolled.

For example, the close button can now be closed with:
  ClickElement(kGlicContentsElementId, kPathToMockGlicCloseButton)

ExecuteJs and EvaluateJs can also be used directly on the glic
contents as well.

Change-Id: Id81eab0d4b8a845f7153cbbaeaef7dba015106e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6190717
Reviewed-by: Dan H <harringtond@chromium.org>
Reviewed-by: Ian Vollick <vollick@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1411021}