glic: Rename method/members of GlicWindowController.
Previously we were going to move to a design where we moved the root
view instead of reparenting the child widget. We've decided to for now
keep reparenting the child widget. Update some nomenclature to reflect
this.

Bug: 391402352
Change-Id: I5f9547e8eb6e97e24b1828928c5724c8a92f9dd2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6198524
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Erik Chen <erikchen@chromium.org>
Reviewed-by: Ian Wells <iwells@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1411098}