Add cc::PropertyTreeDelegate::UpdateScrollOffsetFromImpl.
This adds another entry point to the PropertyTreeDelegate to
manage the differences between layer tree mode and layer list mode
in the cc::LayerTreeHost.

This also changes CompositorPropertyTreeDelegate to be a subclass
of PropertyTreeLayerTreeDelegate to better keep the two files in
sync by default.

This is essentially a refactoring change and should have no
functional effect.

Bug: 389771428
Change-Id: I83d13a2dfcf3ece5a7d9899623c19b33c053f52e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6191551
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1411104}