[TSLO] Handle simultaneously triggered strip fade and height transitions
Currently, updating the strip visibility / scrim opacity during a fade
or height transition is not coordinated in StripLayoutHelperManager.
The issue in the linked bug arises because the visibility is
undesirably updated by the height transition that runs shortly after the
fade transition when both transitions are triggered.

This CL introduces a boolean to determine if height transition
animations that affect the strip scrim should run; in case both fade and
height transitions are triggered simultaneously, the fade transition
will take precedence to update the strip scrim / visibility assuming
that the height transition runs solely to update the strip top padding
and not to show / hide the strip.

Bug: 389710380
Change-Id: Ieebf56fad16eceafabc767da0508e04e89cad500
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6180107
Commit-Queue: Aishwarya Rajesh <aishwaryarj@google.com>
Reviewed-by: Wenyu Fu <wenyufu@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1411109}