[Protected Audience] KVv2 Caching CL 2 of 3
Add an LRU list to TrustedSignalsCacheImpl. This allows caching entries
that are no longer actively in use. Entries are added to the list to
keep them alive when the last Handle keeping them alive is destroyed, they've finished loading, are reusable, and the max cache size is not
currently exceeded.

Live entries may cause the "maximum" size to be exceeded.

This makes for a relatively simple implementation, that neither evicts everything when live cache entries exceed the maximum size,
nor can keep around entries in a way causing the max size to be
exceeded.

In a future CL, entries will be deleted on a timer, and there are
other improvements that could be considered (prioritizing buyer
/ seller signals independently, deleting entries on expiration, etc).

Bug: 333445540
Change-Id: Iad0d8bc2fe58c0cd84af25a94bcb4b39eeec6213
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6179536
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Commit-Queue: mmenke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1411115}