Fix NPEs when a navigation happens while waiting for async interception
Since we've made async navigation interception async in Java, we have
to handle various events like getting backgrounded clearing the
RedirectHandler, or the user switching tabs.

In situations like this, it's probably not a good idea to launch an
app, so we should just cancel the check and have the navigation
proceed.

Bug: 381535042, 391842690, 392131251
Change-Id: I4e5bbdefd30a56b46d4990f2b203596b8b8c8a07
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6196277
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Commit-Queue: Michael Thiessen <mthiesse@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1411151}