Add deletion dialog for shared tab groups
This CL specializes the dialog text when deleting a saved tab group, in
the case where the group is also shared. This specifically covers the
case where the owner is explicitly deleting a group, not when the group
is implicitly deleted by becoming empty.

Low coverage seems to mostly just be how the tool is handling end of
scopes after early returns, it's a false positive

Low-Coverage-Reason: COVERAGE_UNDERREPORTED
Change-Id: I234cb606d84bad71386a86592f3429395e714e80
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6182508
Reviewed-by: Darryl James <dljames@chromium.org>
Commit-Queue: Taylor Bergquist <tbergquist@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1411162}